Schedule and Special Note
I make a schedule for this project in part 1, and we can add note on the part 2, in order to make note for the final project report.
I just try to use LaTex to do something and try to upload files to GitHub. Thank you to let me open new world.